requirements.txt
numpy>=1.24.3
numpy-stl>=3.0.1
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
SQLAlchemy>=2.0.15
Werkzeug>=2.3.4
PyJWT>=2.8.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
fastapi-mail>=1.4.1
pydantic[email]>=2.5.0
